requirements.txt
SpeechRecognition==3.10.4
pyttsx3==2.90
PyAudio==0.2.14
google-generativeai>=0.5.0
python-dotenv==1.0.1
